requirements.txt
streamlit==1.39.0
numpy==1.26.4
pandas==2.2.2
plotly==5.24.1
yfinance==0.2.43
python-docx==1.1.2
docxtpl==0.16.8
